+++ src/commands/mod.rs
@@ -1,6 +1,7 @@
 pub mod clean;
 pub mod clean_all;
 pub mod list;
+pub mod network;
 pub mod run;
 pub mod setup;
 pub mod shell;